[WL21389] Fixes warnings from last commit (details below).
* Fixes int vs size_t comparison warning within ShapeInfo.
* Fixes unused var warning for shapeVerts local var within GeometryCache.

This should fix the mac & ubuntu builds.

Changes Committed:
    modified:   libraries/render-utils/src/GeometryCache.cpp
    modified:   libraries/shared/src/ShapeInfo.cpp

+++ libraries/shared/src/ShapeInfo.cpp
@@ -40,7 +40,7 @@ static const size_t SHAPETYPE_NAME_COUNT = (sizeof(shapeTypeNames) / sizeof((sha
 const float MIN_HALF_EXTENT = 0.005f; // 0.5 cm
 
 QString ShapeInfo::getNameForShapeType(ShapeType type) {
-    if (((int)type <= 0) || ((int)type >= SHAPETYPE_NAME_COUNT)) {
+    if (((int)type <= 0) || ((int)type >= (int)SHAPETYPE_NAME_COUNT)) {
         type = (ShapeType)0;
     }
